+++ sparrow/src/Classes/SPButton.m
@@ -3,7 +3,7 @@
 //  Sparrow
 //
 //  Created by Daniel Sperl on 13.07.09.
-//  Copyright 2011 Gamua. All rights reserved.
+//  Copyright 2011-2014 Gamua. All rights reserved.
 //
 //  This program is free software; you can redistribute it and/or modify
 //  it under the terms of the Simplified BSD License.
@@ -65,6 +65,7 @@
         _enabled = YES;
         _isDown = NO;
         _textBounds = [[SPRectangle alloc] initWithX:0 y:0 width:_upState.width height:_upState.height];
+        self.touchGroup = YES;
         
         [_contents addChild:_background];
         [self addChild:_contents];
@@ -257,6 +258,16 @@
 {
     [self createTextField];
     _textField.color = value;
+}
+
+- (uint)color
+{
+    return _background.color;
+}
+
+- (void)setColor:(uint)color
+{
+    _background.color = color;
 }
 
 - (void)setUpState:(SPTexture *)upState