+++ sparrow/src/Classes/SPDisplayObject.h
@@ -3,7 +3,7 @@
 //  Sparrow
 //
 //  Created by Daniel Sperl on 15.03.09.
-//  Copyright 2011 Gamua. All rights reserved.
+//  Copyright 2011-2014 Gamua. All rights reserved.
 //
 //  This program is free software; you can redistribute it and/or modify
 //  it under the terms of the Simplified BSD License.
@@ -12,6 +12,10 @@
 #import <Foundation/Foundation.h>
 #import <Sparrow/SPEventDispatcher.h>
 #import <Sparrow/SPDisplayObjectExport.h>
+
+#ifdef SP_PHYSICS_CLASS
+@class SP_PHYSICS_CLASS;
+#endif
 
 @class SPDisplayObjectContainer;
 @class SPFragmentFilter;
@@ -129,6 +133,10 @@
 /// The y coordinate of the object's origin in its own coordinate space (default: 0).
 @property (nonatomic, assign) float pivotY;
 
+/// The scale factor. "1" means no scale, a negative value inverts the object.
+/// Note: Accessing this property will always return scaleX, even if scaleX and scaleY are not equal.
+@property (nonatomic, assign) float scale;
+
 /// The horizontal scale factor. "1" means no scale, negative values flip the object.
 @property (nonatomic, assign) float scaleX;
 
@@ -197,6 +205,10 @@
 /// The physics body associated with the display object. Sparrow does not provide physics on its
 /// own, but this property may be used by any physics library to link an object to its physical
 /// body.
+#ifdef SP_PHYSICS_CLASS
+@property (nonatomic, strong) SP_PHYSICS_CLASS *physicsBody;
+#else
 @property (nonatomic, strong) id physicsBody;
+#endif
 
 @end