+++ sparrow/src/Classes/SPDisplayObjectContainer.m
@@ -3,7 +3,7 @@
 //  Sparrow
 //
 //  Created by Daniel Sperl on 15.03.09.
-//  Copyright 2011 Gamua. All rights reserved.
+//  Copyright 2011-2014 Gamua. All rights reserved.
 //
 //  This program is free software; you can redistribute it and/or modify
 //  it under the terms of the Simplified BSD License.
@@ -41,6 +41,7 @@ static void getDescendantEventListeners(SPDisplayObject *object, NSString *event
 @implementation SPDisplayObjectContainer
 {
     NSMutableArray *_children;
+    BOOL _touchGroup;
 }
 
 #pragma mark Initialization
@@ -262,7 +263,9 @@ static void getDescendantEventListeners(SPDisplayObject *object, NSString *event
         SPMatrix *transformationMatrix = [self transformationMatrixToSpace:child];
         SPPoint  *transformedPoint = [transformationMatrix transformPoint:localPoint];
         SPDisplayObject *target = [child hitTestPoint:transformedPoint];
-        if (target) return target;
+
+        if (target)
+            return _touchGroup ? self : target;
     }
 
     return nil;