+++ sparrow/src/Classes/SPMacros.h
@@ -3,7 +3,7 @@
 //  Sparrow
 //
 //  Created by Daniel Sperl on 15.03.09.
-//  Copyright 2011 Gamua. All rights reserved.
+//  Copyright 2011-2014 Gamua. All rights reserved.
 //
 //  This program is free software; you can redistribute it and/or modify
 //  it under the terms of the Simplified BSD License.
@@ -88,6 +88,15 @@ SP_INLINE uint SPHashFloat(float value)
 {
     union { float f; int i; } converter = { .f = value };
     return converter.i & 0xffffff00; // mask for epsilon
+}
+
+SP_INLINE uint SPHashPointer(void *ptr)
+{
+  #ifdef __LP64__
+    return (uint)(((uintptr_t)ptr) >> 3);
+  #else
+    return ((uintptr_t)ptr) >> 2;
+  #endif
 }
 
 SP_INLINE uint SPShiftAndRotate(uint value, int shift)
@@ -178,7 +187,6 @@ SP_EXTERN NSString *const SPExceptionOperationFailed;
 #else
     #define SP_RELEASE_AND_COPY(_old, _new)     \
         do {                                    \
-            if (_old == _new) break;            \
             id tmp = _old;                      \
             _old = [_new copy];                 \
             [tmp release];                      \