+++ sparrow/src/Classes/SPOpenGL.h
@@ -3,6 +3,7 @@
 //  Sparrow
 //
 //  Created by Robert Carone on 10/8/13.
+//  Copyright 2011-2014 Gamua. All rights reserved.
 //
 //  This program is free software; you can redistribute it and/or modify
 //  it under the terms of the Simplified BSD License.
@@ -10,16 +11,43 @@
 
 #import <OpenGLES/ES2/gl.h>
 #import <OpenGLES/ES2/glext.h>
+
 #import <Sparrow/SPMacros.h>
 
-#define SP_ENABLE_GL_STATE_CACHE        1
+// -----------------------------------------------------------
+// EXPERIMENTAL FEATURE: Activate the OpenGL state cache here!
+// -----------------------------------------------------------
+
+#define SP_ENABLE_GL_STATE_CACHE 0
+
+/// Sparrow's OpenGL state cache reference type.
+typedef struct SGLStateCache *SGLStateCacheRef;
+
+/// Allocates a new state cache.
+SP_EXTERN SGLStateCacheRef sglStateCacheCreate(void);
+
+/// Returns a copy of a state cache.
+SP_EXTERN SGLStateCacheRef sglStateCacheCopy(SGLStateCacheRef stateCache);
+
+/// Deallocates a state cache.
+SP_EXTERN void sglStateCacheRelease(SGLStateCacheRef stateCache);
+
+/// Resets the state cache's values to an null state.
+SP_EXTERN void sglStateCacheReset(SGLStateCacheRef stateCache);
+
+/// Returns the current global state cache.
+SP_EXTERN SGLStateCacheRef sglStateCacheGetCurrent(void);
+
+/// Sets the current global state cache, if NULL will use a default state cache.
+SP_EXTERN void sglStateCacheSetCurrent(SGLStateCacheRef stateCache);
 
 /// Returns a string representing an OpenGL error code.
-SP_EXTERN const GLchar*                 sglGetErrorString(GLenum error);
+SP_EXTERN const char* sglGetErrorString(uint error);
 
-// GL_OES_vertex_array_object
-
+/// Extension remappings
 #if GL_OES_vertex_array_object
+    #undef GL_VERTEX_ARRAY_BINDING
+
     #define GL_VERTEX_ARRAY_BINDING     GL_VERTEX_ARRAY_BINDING_OES
     #define glBindVertexArray           glBindVertexArrayOES
     #define glGenVertexArrays           glGenVertexArraysOES
@@ -27,14 +55,10 @@ SP_EXTERN const GLchar*                 sglGetErrorString(GLenum error);
     #define glIsVertexArray             glIsVertexArrayOES
 #endif
 
-// OpenGL state cache
-
+/// OpenGL remappings
 #if SP_ENABLE_GL_STATE_CACHE
-
     #undef  glBindVertexArray
     #undef  glDeleteVertexArrays
-
-    // shims
 
     #define glActiveTexture             sglActiveTexture
     #define glBindBuffer                sglBindBuffer
@@ -56,8 +80,6 @@ SP_EXTERN const GLchar*                 sglGetErrorString(GLenum error);
     #define glUseProgram                sglUseProgram
     #define glViewport                  sglViewport
 
-    // prototypes
-
     SP_EXTERN void                      sglActiveTexture(GLenum texture);
     SP_EXTERN void                      sglBindBuffer(GLenum target, GLuint buffer);
     SP_EXTERN void                      sglBindFramebuffer(GLenum target, GLuint framebuffer);
@@ -77,5 +99,4 @@ SP_EXTERN const GLchar*                 sglGetErrorString(GLenum error);
     SP_EXTERN void                      sglScissor(GLint x, GLint y, GLsizei width, GLsizei height);
     SP_EXTERN void                      sglUseProgram(GLuint program);
     SP_EXTERN void                      sglViewport(GLint x, GLint y, GLsizei width, GLsizei height);
-
-#endif //!SP_ENABLE_GL_STATE_CACHE
+#endif