+++ sparrow/src/Classes/SPSprite.m
@@ -3,7 +3,7 @@
 //  Sparrow
 //
 //  Created by Daniel Sperl on 21.03.09.
-//  Copyright 2011 Gamua. All rights reserved.
+//  Copyright 2011-2014 Gamua. All rights reserved.
 //
 //  This program is free software; you can redistribute it and/or modify
 //  it under the terms of the Simplified BSD License.
@@ -33,6 +33,7 @@
 - (void)dealloc
 {
     [_flattenedContents release];
+    [_clipRect release];
     [super dealloc];
 }