+++ sparrow/src/Classes/SPTouchProcessor.m
@@ -3,7 +3,7 @@
 //  Sparrow
 //
 //  Created by Daniel Sperl on 03.05.09.
-//  Copyright 2011 Gamua. All rights reserved.
+//  Copyright 2011-2014 Gamua. All rights reserved.
 //
 //  This program is free software; you can redistribute it and/or modify
 //  it under the terms of the Simplified BSD License.
@@ -136,6 +136,13 @@
 {
     double now = CACurrentMediaTime();
     
+    // remove touches that have already ended / were already canceled
+    [_currentTouches filterUsingPredicate:
+     [NSPredicate predicateWithBlock:^BOOL(SPTouch *touch, NSDictionary *bindings)
+      {
+          return touch.phase != SPTouchPhaseEnded && touch.phase != SPTouchPhaseCancelled;
+      }]];
+
     for (SPTouch *touch in _currentTouches)
     {
         touch.phase = SPTouchPhaseCancelled;